Moved the border filter to a new namespace
